Move scriptException to js/errors.go
